AI Scripts/requirements.txt
openai
os
langchain
pinecone
python-dotenv
langchain-community
pypdf
langchain_openai